gpt-eval-kit/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gpt-eval-kit"
version = "0.1.0"
description = "OpenAI-first deployment-grade evaluation framework."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.40.0",
  "jsonschema>=4.23.0",
  "PyYAML>=6.0.0",
]

[project.scripts]
evalkit = "evalkit.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["evalkit*"]
